Add support for Piko 15
To keep backwards compatibility with the previous versions of this
script I've added a --piko-model flag which in turn decides which
metrics to scrape (and to allocate for in the database schema).

This flag adds dc inputs to cover up to three inputs.

Besides the dc input no.2 that were not used before there were two
metrics that now are enabled for Piko 15.