Add importer script for LogData.dat
This script can parse the LogData.dat file that can be exported from the
inverters local web site. There are some caveats though; quite low
sample rate and metric count, but it's better than nothing when you
start off with an empty database.

As stated in the script, it's highly possible that I've missed some
metric from the log file, but it is easy to add if needed.